Add ImagenImage and ImagenImageRepresentable
